Fix diag server shutdown, expose ring buffer for tests, clean up LCL logging
- lcldiagserver: Move TLCLEventRing and TLCLDiagServer to interface for testability
- lcldiagserver: Add OnAcceptError handler + StopAccepting(true) for clean shutdown
  (pattern from TFpDebugTcpServer in components/fpdebug/)
- lcldiagserver: Use class cracker for TControl.AutoSizingLockCount access
- control.inc: Gate verbose DebugLn behind {$IFDEF LAZ_LAYOUT_DEBUG}
- controls.pp: Remove unused WMSizeCounter global
- forms.pp + application.inc: Hook LCLDiagServer startup into TApplication.Run
- build.sh: Enable DIAG=1 by default, replace DEBUG_WM_SIZE with ENABLE_LCL_SOCKET_DIAG
- calcform.pas: Wrap BeginFormUpdate/EndFormUpdate in try..finally
- customform.inc, customdockform.inc, wincontrol.inc: Clean up exception handling
- aboutfrm.pas: Bold + white font styling for all About dialog labels
- lazarusidestrconsts.pas: Fix double-percent escape in format string
- Docs: Clarify GNOME/Mutter vs KDE Plasma resize behavior
- Tests: 25 ring buffer + 19 server socket tests (44 total, all passing)
- Tests: Add testserversimple.lpr standalone diagnostic

+++ test/lcltests/testdiagserver.pas
@@ -218,7 +218,7 @@ begin
     Resp := SendRecv(Sock, '{"id":1,"cmd":"ping"}');
     Pid := ExtractJInt(Resp, 'pid');
     AssertTrue('pid is positive', Pid > 0);
-    AssertEquals('pid matches current process', GetProcessID, Pid);
+    AssertEquals('pid matches current process', Int64(GetProcessID), Pid);
   finally
     Sock.Free;
   end;
